Improve package handling and useability
Vendor metadata now contain variable , ,  and  to make it easier to change module metadata. Every module now contain an variable  which ensures that the folder name and module ID are equal. Also the variables  and  can represent the path to this module thus it makes easier to create extends paths for example: / = package_name/module_name